+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This file is used to list changes made in each version of the Java cookbook.
 - Remove platform suport for untested platforms (smartOS, XenServer, zlinux, arch)
 - Remove Ubuntu 14.04 support
 - Fixed oracle download link for JDK 8 (update to 8u202 from 8u201)
+- fixed specs for windows
 
 ## 3.2.0 - (2019-01-24)